Split out BufferManagerEnqueue and rename to BufferEnqueue.
